# docs/swagger.yaml
basePath: /api
definitions:
  dto.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UpsertUserDto:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.CommentModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  models.PhotoModel:
    properties:
      comment:
        items:
          $ref: '#/definitions/models.CommentModel'
        type: array
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      photo_id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  models.SocialMediaModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      social_media_id:
        type: integer
      social_media_url:
        type: string
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  models.UserModel:
    properties:
      age:
        type: integer
      comments:
        items:
          $ref: '#/definitions/models.CommentModel'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.PhotoModel'
        type: array
      social_medias:
        items:
          $ref: '#/definitions/models.SocialMediaModel'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
  utils.HttpError:
    properties:
      message: {}
      stack_trace: {}
    type: object
  utils.HttpSuccess-array_models_UserModel:
    properties:
      data:
        items:
          $ref: '#/definitions/models.UserModel'
        type: array
      message: {}
    type: object
  utils.HttpSuccess-dto_UpsertUserDto:
    properties:
      data:
        $ref: '#/definitions/dto.UpsertUserDto'
      message: {}
    type: object
  utils.HttpSuccess-models_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/models.LoginResponse'
      message: {}
    type: object
host: localhost:3030
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /user:
    delete:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-dto_UpsertUserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: delete current user based on JWT
      tags:
      - User
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-array_models_UserModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: get mutilple users
      tags:
      - User
    put:
      parameters:
      - description: Update User Based On Token
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertUserDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-dto_UpsertUserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: create a user
      tags:
      - User
  /user/login:
    post:
      parameters:
      - description: Login User DTO
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-models_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: login a user
      tags:
      - User
  /user/register:
    post:
      parameters:
      - description: Create User DTO
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertUserDto'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.HttpSuccess-dto_UpsertUserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: create a user
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"